simba/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# Launch_tkinter_GUI.py: 32
# extract_seqframes.py: 7
# labelling_aggression.py: 3
# next_greyscale.py: 7
# next_superimposeframe.py: 7
# raw_scripts\labelling_aggression.py: 3
# tkinter_functions.py: 8
Pillow == 5.4.1

# Launch_tkinter_GUI.py: 5
# dlc_change_yamlfile.py: 5
deeplabcut == 2.0.5

# train_model_2.py: 18,19
eli5 == 0.10.1

# raw_scripts\sklearn_DLC_RF_train_model.py: 16
# sklearn_DLC_RF_train_model.py: 16
# train_model_2.py: 9,10
# train_multiple_models_from_meta.py: 7,8
imblearn == 0.0

# path_plot.py: 10
# raw_scripts\path_plot.py: 11
imutils == 0.5.2

# gantt.py: 2
# line_plot.py: 5
# raw_scripts\gantt.py: 2
# raw_scripts\line_plot.py: 5
# raw_scripts\sklearn_DLC_RF_train_model.py: 14
# sklearn_DLC_RF_train_model.py: 14
# train_model_2.py: 20
# train_multiple_models_from_meta.py: 12
matplotlib == 2.2.2

# correct_devs_loc.py: 4
# correct_devs_mov.py: 3
# data_plot.py: 5
# extract_features_with_scaling.py: 6
# extract_features_wo_targets.py: 6
# extract_seqframes.py: 10
# gantt.py: 4
# get_coordinates_tools_v2.py: 4
# line_plot.py: 3
# merge_frames.py: 3
# next_clahe.py: 3
# path_plot.py: 5
# plot_sklearn_results.py: 1
# process_data_log.py: 3
# process_movement.py: 7
# process_severity.py: 4
# raw_scripts\Run_RF_model.py: 4
# raw_scripts\correct_devs_loc.py: 4
# raw_scripts\correct_devs_mov.py: 3
# raw_scripts\data_plot.py: 5
# raw_scripts\extract_features_old.py: 6
# raw_scripts\extract_features_with_scaling.py: 6
# raw_scripts\gantt.py: 4
# raw_scripts\get_coordinates.py: 4
# raw_scripts\get_coordinates_2.py: 4
# raw_scripts\line_plot.py: 3
# raw_scripts\merge_frames.py: 3
# raw_scripts\path_plot.py: 6
# raw_scripts\plot_sklearn_results.py: 1
# raw_scripts\process_data_log.py: 3
# raw_scripts\process_movement.py: 7
# raw_scripts\process_severity.py: 4
# run_RF_model.py: 4
# tkinter_functions.py: 2
# train_model_2.py: 17
numpy == 1.14.6

# classifier_validation.py: 5
# data_plot.py: 6
# dlc_change_yamlfile.py: 2
# get_coordinates_tools_v2.py: 3
# import_videos_csv_project_ini.py: 3
# merge_frames.py: 2
# merge_movie_ffmpeg.py: 3
# next_clahe.py: 1
# next_greyscale.py: 3
# next_superimposeframe.py: 3
# path_plot.py: 2
# play_video.py: 1
# plot_sklearn_results.py: 2
# process_videos_automation.py: 2
# raw_scripts\data_plot.py: 6
# raw_scripts\get_coordinates.py: 3
# raw_scripts\get_coordinates_2.py: 3
# raw_scripts\merge_frames.py: 2
# raw_scripts\merge_movie_ffmpeg.py: 3
# raw_scripts\merge_movie_opencv.py: 1
# raw_scripts\path_plot.py: 3
# raw_scripts\plot_sklearn_results.py: 2
# tkinter_functions.py: 4
opencv_python == 3.4.5.20

# correct_devs_loc.py: 7,8
# correct_devs_mov.py: 9,10
# data_plot.py: 8,9
# gantt.py: 6,7
# raw_scripts\data_plot.py: 8,9
# raw_scripts\gantt.py: 6,7
openpyxl == 2.6.2

# classifier_validation.py: 2
# correct_devs_loc.py: 1
# correct_devs_mov.py: 1
# data_plot.py: 2
# extract_features_with_scaling.py: 3
# extract_features_wo_targets.py: 3
# gantt.py: 1
# labelling_aggression.py: 2
# line_plot.py: 2
# merge_movie_ffmpeg.py: 7
# path_plot.py: 3
# play_video.py: 5
# plot_sklearn_results.py: 4
# process_data_log.py: 1
# process_movement.py: 1
# process_severity.py: 1
# raw_scripts\Run_RF_model.py: 1
# raw_scripts\correct_devs_loc.py: 1
# raw_scripts\correct_devs_mov.py: 1
# raw_scripts\data_plot.py: 2
# raw_scripts\extract_features_old.py: 3
# raw_scripts\extract_features_with_scaling.py: 3
# raw_scripts\gantt.py: 1
# raw_scripts\get_coordinates.py: 5
# raw_scripts\get_coordinates_2.py: 5
# raw_scripts\labelling_aggression.py: 2
# raw_scripts\line_plot.py: 2
# raw_scripts\merge_movie_ffmpeg.py: 7
# raw_scripts\path_plot.py: 4
# raw_scripts\plot_sklearn_results.py: 4
# raw_scripts\process_data_log.py: 1
# raw_scripts\process_movement.py: 1
# raw_scripts\process_severity.py: 1
# raw_scripts\sklearn_DLC_RF_train_model.py: 1
# raw_scripts\target_insert.py: 1
# run_RF_model.py: 1
# sklearn_DLC_RF_train_model.py: 1
# train_model_2.py: 3
# train_multiple_models_from_meta.py: 3
pandas == 0.21.0

# extract_features_with_scaling.py: 7,8
# extract_features_wo_targets.py: 7,8
# raw_scripts\extract_features_old.py: 7,8
# raw_scripts\extract_features_with_scaling.py: 7,8
# raw_scripts\path_plot.py: 10
# raw_scripts\plot_sklearn_results.py: 7
scipy == 1.1.0

# path_plot.py: 9
# raw_scripts\path_plot.py: 2
seaborn == 0.9.0

# raw_scripts\Run_RF_model.py: 3
# raw_scripts\sklearn_DLC_RF_train_model.py: 2,3,4,5,6,11
# run_RF_model.py: 3
# sklearn_DLC_RF_train_model.py: 2,3,4,5,6,11
# train_model_2.py: 23
# train_multiple_models_from_meta.py: 4,5,9,10
sklearn == 0.0

# train_multiple_models_from_meta.py: 15
tabulate == 0.8.3

# raw_scripts\merge_movie_opencv.py: 2
tqdm == 4.30.0

# raw_scripts\sklearn_DLC_RF_train_model.py: 15
# sklearn_DLC_RF_train_model.py: 15
# train_model_2.py: 8
xgboost == 0.90

# raw_scripts\sklearn_DLC_RF_train_model.py: 13
# sklearn_DLC_RF_train_model.py: 13
# train_model_2.py: 7
# train_multiple_models_from_meta.py: 6
yellowbrick == 0.9.1
